refactor: remove UserDataRepository and Impl in src/app/core/user
