+++ src/app/core/user/data/UserEntity.ts
