+++ src/app/di/AppDiContainer.ts
@@ -1,7 +1,7 @@
 import { AuthRepositoryImpl } from "../modules/Auth/data/user_credentials/AuthRepositoryImpl";
 import { AuthRepository } from "../modules/Auth/domain/user_credentials/AuthRepository";
-import { UserDataRepositoryImpl } from "../modules/repositories/user_data_repository/data/UserDataRepositoryImpl";
-import { UserDataRepository } from "../modules/repositories/user_data_repository/domain/UserDataRepository";
+import { UserDataRepositoryImpl } from "../core/user/data/UserDataRepositoryImpl";
+import { UserDataRepository } from "../core/user/domain/UserDataRepository";
 
 import { DiContainer } from "./DiContainer";
