+++ src/app/di/DiContainer.ts
@@ -1,5 +1,5 @@
 import { AuthRepository } from "../modules/Auth/domain/user_credentials/AuthRepository";
-import { UserDataRepository } from "../modules/repositories/user_data_repository/domain/UserDataRepository";
+import { UserDataRepository } from "../core/user/domain/UserDataRepository";
 
 export interface DiContainer {
   getAuthRepository(): AuthRepository;
